+++ angular-app/src/app/app.component.html
@@ -13,4 +13,5 @@
     <input name="link">
   </div>
 </form>
+<app-todo-form></app-todo-form>
