+++ angular-app/src/app/todo-form/todo-form.component.ts
@@ -1,0 +1,15 @@
+import { Component, OnInit } from '@angular/core';
+
+@Component({
+  selector: 'app-todo-form',
+  templateUrl: './todo-form.component.html',
+  styleUrls: ['./todo-form.component.css']
+})
+export class TodoFormComponent implements OnInit {
+
+  constructor() { }
+
+  ngOnInit() {
+  }
+
+}
